Move test_dist_hybrid_index_drop_after_sctx_allocation to a new file test_hybrid_dist.py
